much more efficient caching

+++ src/modules/seperator.c
@@ -4,7 +4,7 @@ void ffPrintSeperator(FFinstance* instance)
 {
     ffPrintLogoLine(instance);
 
-    for(uint8_t i = 0; i < instance->config.titleLength; i++)
+    for(uint32_t i = 0; i < instance->config.titleLength; i++)
         putchar('-');
     putchar('\n');
 }
